app/requirements.txt
opencv-python
numpy
onnxruntime
fastapi
uvicorn
python-multipart
Pillow